backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.1
geopy==2.4.1
gunicorn==21.2.0
